Move TotalPlaytimeCounter to module Player
